fix: 补充资源回退获取路径 (#2168)
- Yaml读取回退改到YamlConfig级
- RuntimeLauncher打包修复文件夹名称
- Installer不再从MEIPASS复制资源到工作目录

+++ src/one_dragon/base/config/yaml_operator.py
@@ -1,6 +1,4 @@
 import os
-import sys
-from typing import Optional
 
 import yaml
 
@@ -8,18 +6,6 @@ from one_dragon.utils import yaml_utils
 from one_dragon.utils.log_utils import log
 
 cached_yaml_data: dict[str, tuple[float, dict]] = {}
-
-
-def get_temp_config_path(file_path: str) -> str:
-    """
-    优先检查PyInstaller运行时的_MEIPASS目录下是否有对应的yml文件
-    有则返回该路径，否则返回原路径
-    """
-    if hasattr(sys, '_MEIPASS'):
-        mei_path = os.path.join(sys._MEIPASS, 'config', os.path.basename(file_path))
-        if os.path.exists(mei_path):
-            return mei_path
-    return file_path
 
 def read_cache_or_load(file_path: str):
     cached = cached_yaml_data.get(file_path)
@@ -36,13 +22,13 @@ def read_cache_or_load(file_path: str):
 
 class YamlOperator:
 
-    def __init__(self, file_path: Optional[str] = None):
+    def __init__(self, file_path: str | None = None):
         """
         yml文件的操作器
         :param file_path: yml文件的路径。不传入时认为是mock，用于测试。
         """
 
-        self.file_path: str = get_temp_config_path(file_path) if file_path else None
+        self.file_path: str | None = file_path
         """yml文件的路径"""
 
         self.data: dict = {}
@@ -105,6 +91,8 @@ class YamlOperator:
         删除配置文件
         :return:
         """
+        if self.file_path is None:
+            return
         if os.path.exists(self.file_path):
             os.remove(self.file_path)
